+++ pkg/provider/oidc.go
@@ -17,8 +17,6 @@ const (
 	oidcExtraDataFieldIDToken       = "id_token"
 	oidcExtraDataFieldIDTokenClaims = "id_token_claims"
 	oidcExtraDataFieldUserInfo      = "user_info"
-
-	oidcExchangeConfigOptionNonce = "nonce"
 )
 
 var (
@@ -30,37 +28,36 @@ func init() {
 	GlobalRegistry.MustRegister("oidc", OIDCFactory)
 }
 
-type oidcExchangeConfig struct {
-	delegate        *basicExchangeConfig
+type oidcOperations struct {
+	*basicOperations
 	p               *gooidc.Provider
-	nonce           string
 	extraDataFields []string
 }
 
-func (c *oidcExchangeConfig) verifyUpdateToken(ctx context.Context, t *Token) error {
+func (oo *oidcOperations) verifyUpdateToken(ctx context.Context, t *Token, nonce string) error {
 	rawIDToken, ok := t.Extra("id_token").(string)
 	if !ok {
 		return ErrOIDCMissingIDToken
 	}
 
-	idToken, err := c.p.Verifier(&gooidc.Config{ClientID: c.delegate.config.ClientID}).Verify(ctx, rawIDToken)
+	idToken, err := oo.p.Verifier(&gooidc.Config{ClientID: oo.basicOperations.base.ClientID}).Verify(ctx, rawIDToken)
 	if err != nil {
 		return fmt.Errorf("oidc: verification error: %w", err)
 	}
 
-	// If nonce is configured, make sure it matches the nonce in
-	// the ID token.  It is not configured when refresh_token is
-	// sent in from an external source.
-	if len(c.nonce) > 0 &&
-		(subtle.ConstantTimeEq(int32(len(idToken.Nonce)), int32(len(c.nonce))) == 0 ||
-			subtle.ConstantTimeCompare([]byte(idToken.Nonce), []byte(c.nonce)) == 0) {
+	// If nonce is configured, make sure it matches the nonce in the ID token.
+	// It is not configured when refresh_token is sent in from an external
+	// source.
+	if nonce != "" &&
+		(subtle.ConstantTimeEq(int32(len(idToken.Nonce)), int32(len(nonce))) == 0 ||
+			subtle.ConstantTimeCompare([]byte(idToken.Nonce), []byte(nonce)) == 0) {
 		return ErrOIDCNonceMismatch
 	}
 
-	if len(c.extraDataFields) > 0 {
+	if len(oo.extraDataFields) > 0 {
 		t.ExtraData = make(map[string]interface{})
 
-		for _, field := range c.extraDataFields {
+		for _, field := range oo.extraDataFields {
 			switch field {
 			case oidcExtraDataFieldIDToken:
 				t.ExtraData[field] = rawIDToken
@@ -72,7 +69,7 @@ func (c *oidcExchangeConfig) verifyUpdateToken(ctx context.Context, t *Token) er
 
 				t.ExtraData[field] = claims
 			case oidcExtraDataFieldUserInfo:
-				userInfo, err := c.p.UserInfo(ctx, c.delegate.config.TokenSource(ctx, t.Token))
+				userInfo, err := oo.p.UserInfo(ctx, oo.basicOperations.base.TokenSource(ctx, t.Token))
 				if err != nil {
 					return fmt.Errorf("oidc: error fetching user info: %w", err)
 				}
@@ -90,60 +87,36 @@ func (c *oidcExchangeConfig) verifyUpdateToken(ctx context.Context, t *Token) er
 	return nil
 }
 
-func (c *oidcExchangeConfig) Exchange(ctx context.Context, code string, opts ...oauth2.AuthCodeOption) (*Token, error) {
-	t, err := c.delegate.Exchange(ctx, code, opts...)
+func (oo *oidcOperations) AuthCodeExchange(ctx context.Context, code string, opts ...AuthCodeExchangeOption) (*Token, error) {
+	o := &AuthCodeExchangeOptions{}
+	o.ApplyOptions(opts)
+
+	t, err := oo.basicOperations.AuthCodeExchange(ctx, code, opts...)
 	if err != nil {
 		return nil, err
 	}
 
-	if err := c.verifyUpdateToken(ctx, t); err != nil {
+	if err := oo.verifyUpdateToken(ctx, t, o.ProviderOptions["nonce"]); err != nil {
 		return nil, errmark.MarkUser(err)
 	}
 
 	return t, nil
 }
 
-func (c *oidcExchangeConfig) Refresh(ctx context.Context, t *Token) (*Token, error) {
-	t, err := c.delegate.Refresh(ctx, t)
+func (oo *oidcOperations) RefreshToken(ctx context.Context, t *Token, opts ...RefreshTokenOption) (*Token, error) {
+	o := &RefreshTokenOptions{}
+	o.ApplyOptions(opts)
+
+	t, err := oo.basicOperations.RefreshToken(ctx, t, opts...)
 	if err != nil {
 		return nil, err
 	}
 
-	if err := c.verifyUpdateToken(ctx, t); err != nil {
+	if err := oo.verifyUpdateToken(ctx, t, o.ProviderOptions["nonce"]); err != nil {
 		return nil, errmark.MarkUser(err)
 	}
 
 	return t, nil
-}
-
-type oidcExchangeConfigBuilder struct {
-	config          *oauth2.Config
-	p               *gooidc.Provider
-	nonce           string
-	extraDataFields []string
-}
-
-func (cb *oidcExchangeConfigBuilder) WithOption(name, value string) ExchangeConfigBuilder {
-	if name == oidcExchangeConfigOptionNonce {
-		cb.nonce = value
-	}
-
-	return cb
-}
-
-func (cb *oidcExchangeConfigBuilder) WithRedirectURL(redirectURL string) ExchangeConfigBuilder {
-	cb.config.RedirectURL = redirectURL
-	return cb
-}
-
-func (cb *oidcExchangeConfigBuilder) Build() ExchangeConfig {
-	return &oidcExchangeConfig{
-		delegate: &basicExchangeConfig{
-			config: cb.config,
-		},
-		p:               cb.p,
-		extraDataFields: cb.extraDataFields,
-	}
 }
 
 type oidc struct {
@@ -163,16 +136,18 @@ func (o *oidc) Version() int {
 	return o.vsn
 }
 
-func (o *oidc) NewAuthCodeURLConfigBuilder(clientID string) AuthCodeURLConfigBuilder {
-	return NewConformingAuthCodeURLConfigBuilder(o.endpoint(), clientID)
+func (o *oidc) Public(clientID string) PublicOperations {
+	return o.Private(clientID, "")
 }
 
-func (o *oidc) NewExchangeConfigBuilder(clientID, clientSecret string) ExchangeConfigBuilder {
-	return &oidcExchangeConfigBuilder{
-		config: &oauth2.Config{
-			Endpoint:     o.endpoint(),
-			ClientID:     clientID,
-			ClientSecret: clientSecret,
+func (o *oidc) Private(clientID, clientSecret string) PrivateOperations {
+	return &oidcOperations{
+		basicOperations: &basicOperations{
+			base: &oauth2.Config{
+				Endpoint:     o.endpoint(),
+				ClientID:     clientID,
+				ClientSecret: clientSecret,
+			},
 		},
 		p:               o.p,
 		extraDataFields: o.extraDataFields,